Enhanced package structure and divided controller and service
