+++ backend/src/main/java/com/aadeshandreas/ailearning/ai_learning_companion/controller/DocumentController.java
@@ -1,6 +1,8 @@
 package com.aadeshandreas.ailearning.ai_learning_companion.controller;
 
-import com.aadeshandreas.ailearning.ai_learning_companion.model.*;
+import com.aadeshandreas.ailearning.ai_learning_companion.model.coding.*;
+import com.aadeshandreas.ailearning.ai_learning_companion.model.common.ApiResponse;
+import com.aadeshandreas.ailearning.ai_learning_companion.model.common.DocumentText;
 import com.aadeshandreas.ailearning.ai_learning_companion.service.DocumentService;
 import jakarta.validation.Valid;
 import org.slf4j.Logger;
@@ -79,109 +81,6 @@ public class DocumentController {
             ApiResponse<Void> errorResponse = new ApiResponse<>(
                     "Unable to process text",
                     "TEXT_UPLOAD_ERROR",
-                    null
-            );
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
-        }
-    }
-
-    /**
-     * Generates a structured summary from the document previously uploaded in the user's session.
-     *
-     * @return A {@link ResponseEntity} wrapping a generic {@link ApiResponse}. On success,
-     * the ApiResponse's data field will contain a {@link Summary} object. On failure,
-     * it will contain an error message and code with null data.
-     */
-    @GetMapping(value = "/summary")
-    public ResponseEntity<ApiResponse<?>> uploadAndSummarize() {
-        try {
-            Summary summary = documentService.generateSummary();
-            ApiResponse<Summary> successResponse = new ApiResponse<>("Success", "200_OK", summary);
-            return ResponseEntity.ok(successResponse);
-        } catch (IllegalStateException e) {
-            logger.warn("Summary generation failed because no document was uploaded: {}", e.getMessage());
-
-            ApiResponse<Void> errorResponse = new ApiResponse<>(
-                    "No document found. Please upload a document first.",
-                    "NO_DOCUMENT_UPLOADED",
-                    null
-            );
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
-
-        } catch (Exception e) {
-            logger.error("An unexpected error occurred during summary generation", e);
-
-            ApiResponse<Void> errorResponse = new ApiResponse<>(
-                    "Unable to process document",
-                    "INTERNAL_SERVER_ERROR",
-                    null
-            );
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
-        }
-    }
-
-    /**
-     * Generates a list of flashcards from the document previously uploaded in the user's session.
-     *
-     * @return A {@link ResponseEntity} wrapping a generic {@link ApiResponse}. On success,
-     * the ApiResponse's data field will contain a {@link FlashcardList} object. On failure,
-     * it will contain an error message and code with null data.
-     */
-    @GetMapping(value = "/flashcards")
-    public ResponseEntity<ApiResponse<?>> uploadAndGenerateFlashcard() {
-        try {
-            FlashcardList flashcardList = documentService.generateFlashcards();
-            ApiResponse<FlashcardList> successResponse = new ApiResponse<>("Success", "200_OK", flashcardList);
-            return ResponseEntity.ok(successResponse);
-        } catch (IllegalStateException e) {
-            logger.warn("Flashcard generation failed because no document was uploaded: {}", e.getMessage());
-
-            ApiResponse<Void> errorResponse = new ApiResponse<>(
-                    "No document found. Please upload a document first.",
-                    "NO_DOCUMENT_UPLOADED",
-                    null
-            );
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
-        } catch (Exception e) {
-            logger.error("Error processing uploaded PDF", e);
-
-            ApiResponse<Void> errorResponse = new ApiResponse<>(
-                    "Unable to process document",
-                    "INTERNAL_SERVER_ERROR",
-                    null
-            );
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
-        }
-    }
-
-    /**
-     * Generates a quiz from the document previously uploaded in the user's session.
-     *
-     * @return A {@link ResponseEntity} wrapping a generic {@link ApiResponse}. On success,
-     * the ApiResponse's data field will contain a {@link Quiz} object. On failure,
-     * it will contain an error message and code with null data.
-     */
-    @GetMapping(value = "/quiz")
-    public ResponseEntity<ApiResponse<?>> uploadAndGenerateQuiz() {
-        try {
-            Quiz q = documentService.generateQuiz();
-            ApiResponse<Quiz> successResponse = new ApiResponse<>("Success", "200_OK", q);
-            return ResponseEntity.ok(successResponse);
-        } catch (IllegalStateException e) {
-            logger.warn("Quiz generation failed because no document was uploaded: {}", e.getMessage());
-
-            ApiResponse<Void> errorResponse = new ApiResponse<>(
-                    "No document found. Please upload a document first.",
-                    "NO_DOCUMENT_UPLOADED",
-                    null
-            );
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
-        } catch (Exception e) {
-            logger.error("Error processing uploaded PDF", e);
-
-            ApiResponse<Void> errorResponse = new ApiResponse<>(
-                    "Unable to process document",
-                    "INTERNAL_SERVER_ERROR",
                     null
             );
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
