+++ backend/src/main/java/com/aadeshandreas/ailearning/ai_learning_companion/service/generation/FlashcardGenerator.java
@@ -1,6 +1,6 @@
-package com.aadeshandreas.ailearning.ai_learning_companion.service;
+package com.aadeshandreas.ailearning.ai_learning_companion.service.generation;
 
-import com.aadeshandreas.ailearning.ai_learning_companion.model.FlashcardList;
+import com.aadeshandreas.ailearning.ai_learning_companion.model.content.FlashcardList;
 import dev.langchain4j.service.UserMessage;
 import dev.langchain4j.service.V;
