+++ backend/src/main/java/com/aadeshandreas/ailearning/ai_learning_companion/service/generation/QuizGenerator.java
@@ -1,7 +1,6 @@
-package com.aadeshandreas.ailearning.ai_learning_companion.service;
+package com.aadeshandreas.ailearning.ai_learning_companion.service.generation;
 
-import com.aadeshandreas.ailearning.ai_learning_companion.model.FlashcardList;
-import com.aadeshandreas.ailearning.ai_learning_companion.model.Quiz;
+import com.aadeshandreas.ailearning.ai_learning_companion.model.content.Quiz;
 import dev.langchain4j.service.UserMessage;
 import dev.langchain4j.service.V;
